+++ pysui/sui/sui_pgql/pgql_fragments.py
@@ -62,12 +62,39 @@ class StandardCoin(PGQL_Fragment):
                         previous_transaction=schema.TransactionBlock.digest
                     ),
                     schema.Coin.owner.select(
+                        # DSLInlineFragment()
+                        # .on(schema.AddressOwner)
+                        # .select(
+                        #     schema.AddressOwner.owner.select(
+                        #         coin_owner=schema.Owner.address
+                        #     )
+                        # ),
                         DSLInlineFragment()
                         .on(schema.AddressOwner)
                         .select(
                             schema.AddressOwner.owner.select(
-                                coin_owner=schema.Owner.address
-                            )
+                                address_id=schema.Owner.address
+                            ),
+                            obj_owner_kind=DSLMetaField("__typename"),
+                        ),
+                        DSLInlineFragment()
+                        .on(schema.Shared)
+                        .select(
+                            initial_version=schema.Shared.initialSharedVersion,
+                            obj_owner_kind=DSLMetaField("__typename"),
+                        ),
+                        DSLInlineFragment()
+                        .on(schema.Immutable)
+                        .select(
+                            obj_owner_kind=DSLMetaField("__typename"),
+                        ),
+                        DSLInlineFragment()
+                        .on(schema.Parent)
+                        .select(
+                            schema.Parent.parent.select(
+                                parent_id=schema.Object.address
+                            ),
+                            obj_owner_kind=DSLMetaField("__typename"),
                         ),
                     ),
                     schema.Coin.contents.select(